+++ lib/test_prof/event_prof.rb
@@ -186,5 +186,5 @@ module TestProf
 end
 
 require "test_prof/event_prof/rspec" if defined?(RSpec::Core)
-require "test_prof/event_prof/minitest" if defined?(Minitest::Reporters)
+require "test_prof/event_prof/minitest" if defined?(Minitest)
 require "test_prof/event_prof/custom_events"